Button wakeup and database log
Fixed where the call to put the esp back to sleep is made. Now it gives time to the sqlite handler to log a new access, if the tag detected is in the system, before going back to sleep.

Modified logic to wakeup the system with a physical button. Couldn't make it work with two sources of wakeup (ext1 and timer), so instead I made a check for the button state when it wakes up from the timer, which sets a flag to stay awake if the button is low.