Prevent database logs if time not set.
Added a check to prevent any writing of new data if the system just started or failed, in which case the date would be defaulted to 01/01/70 given how the ESP32Time library operates. It only allows new data writing if the simulated rtc is updated.

Added a pinout to allow measurements on how long the esp module stays awake. (GPIO 12).

Fixed csv generation to run asyncronous to the main software, using FreeRTOS and setting the task to core 1. (If the database was too big, the getDB() function would run for too long, triggerig a reset via watchdog).